'alpha': args.alpha,

+++ TrainAndTest/Spectrogram/train_lores10_kaldi.py
@@ -106,7 +106,7 @@ parser.add_argument('--channels', default='64,128,256', type=str,
                     metavar='CHA', help='The channels of convs layers)')
 parser.add_argument('--feat-dim', default=161, type=int, metavar='FEAT',
                     help='acoustic feature dimension')
-parser.add_argument('--alpha', default=12, type=int, metavar='FEAT',
+parser.add_argument('--alpha', default=12, type=float, metavar='FEAT',
                     help='acoustic feature dimension')
 parser.add_argument('--kernel-size', default='5,5', type=str, metavar='KE',
                     help='kernel size of conv filters')
